+++ kiosk_teamProject/kiosk_teamProject/Frozens.swift
@@ -40,7 +40,6 @@ struct Frozens {
                 print("")
             case 0 :
                 print("뒤로가기")
-                Kiosk().shakeShack()
             default:
                 print("다시 선택해주세요")
                 continue
