+++ kiosk_teamProject/kiosk_teamProject/beers.swift
@@ -37,7 +37,6 @@ struct Beers {
                 print("")
             case 0 :
                 print("뒤로가기")
-                Kiosk().shakeShack()
             default:
                 print("다시 선택해주세요")
                 continue
